# frontend/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/webhooks/{id}": {
      "get": {
        "description": "Fetches configuration details for a specific webhook endpoint.",
        "operationId": "getWebhookById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get webhook details",
        "tags": [
          "Webhooks"
        ]
      },
      "patch": {
        "description": "Modifies settings for an existing webhook endpoint, such as the target URL or active status.",
        "operationId": "updateWebhook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update webhook configuration",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
    "/webhook-dlq": {
      "get": {
        "description": "Retrieves a paginated list of failed webhook deliveries currently in the Dead Letter Queue.",
        "operationId": "listWebhookDlq",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by start date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by end date (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookEventId",
            "required": false,
            "in": "query",
            "description": "Filter by specific webhook event ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "required": false,
            "in": "query",
            "description": "Filter by specific webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List failed deliveries",
        "tags": [
          "Webhook DLQ"
        ]
      }
    },
    "/webhook-dlq/{id}/replay": {
      "post": {
        "description": "Manually triggers a retry for a specific failed webhook delivery from the Dead Letter Queue.",
        "operationId": "replayWebhookDlq",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayWebhookDlqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry failed delivery",
        "tags": [
          "Webhook DLQ"
        ]
      }
    },
    "/webhook-dlq/{id}": {
      "get": {
        "description": "Fetches details of a specific failed delivery record from the Dead Letter Queue.",
        "operationId": "getWebhookDlqById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get DLQ item details",
        "tags": [
          "Webhook DLQ"
        ]
      }
    }
  },
  "info": {
    "title": "Hook Relay API",
    "description": "Hook Relay API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/path"
          },
          "message": {
            "example": "ErrorMessage",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "ErrorType"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "message",
          "error"
        ]
      },
...
      "WebhookResponseDto": {
        "type": "object",
        "properties": {
          "webhookId": {
            "type": "string",
            "description": "Unique identifier for the webhook"
          },
          "clientId": {
            "type": "string",
            "description": "ID of the client owning this webhook"
          },
          "endpointName": {
            "type": "string",
            "description": "Descriptive name for the endpoint"
          },
          "targetUrl": {
            "type": "string",
            "description": "The URL to send the webhook payload to"
          },
          "hmacSecret": {
            "type": "object",
            "description": "Secret for HMAC signature verification",
            "nullable": true
          },
          "retryPolicy": {
            "type": "object",
            "description": "Retry policy configuration",
            "nullable": true
          },
          "timeoutMs": {
            "type": "object",
            "description": "Request timeout in milliseconds",
            "default": 5000,
            "nullable": true
          },
          "isActive": {
            "type": "object",
            "description": "Whether the webhook is active",
            "default": false,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "webhookId",
          "clientId",
          "endpointName",
          "targetUrl",
          "timeoutMs",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateWebhookDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "The ID of the client owning this webhook"
          },
          "endpointName": {
            "type": "string",
            "description": "A descriptive name for the endpoint (e.g. 'Order Created')"
          },
          "targetUrl": {
            "type": "string",
            "description": "The URL to send the webhook payload to"
          },
          "hmacSecret": {
            "type": "string",
            "description": "Secret for HMAC signature verification"
          },
          "retryPolicy": {
            "type": "object",
            "description": "Retry policy configuration"
          },
          "timeoutMs": {
            "type": "number",
            "description": "Request timeout in milliseconds",
            "default": 5000
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the webhook is active",
            "default": true
          }
        }
      },
...
      "ReplayWebhookDlqDto": {
        "type": "object",
        "properties": {
          "initiatorId": {
            "type": "string",
            "description": "ID of the user initiating the replay"
          },
          "initiatorType": {
            "type": "string",
            "description": "Type of initiator (system or user)",
            "default": "user"
          }
        }
      }
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}